Agregado el nombre del usuario en la partida y el poder jugar sin registrarse ni iniciar sesion

+++ Entregable_3_1/src/main/java/modelo/Partida.java
@@ -6,10 +6,10 @@ import java.time.LocalDateTime;
 
 public class Partida implements Serializable {
     private final LocalDateTime fecha;
-    private final int tiempoEnSegundos;
+    private final long tiempoEnSegundos;
     private final String jugador;
 
-    public Partida(String jugador, int tiempoEnSegundos) {
+    public Partida(String jugador, long tiempoEnSegundos) {
         this.fecha = LocalDateTime.now();
         this.tiempoEnSegundos = tiempoEnSegundos;
         this.jugador = jugador;
@@ -23,7 +23,7 @@ public class Partida implements Serializable {
         return fecha;
     }
 
-    public int getTiempoEnSegundos() {
+    public long getTiempoEnSegundos() {
         return tiempoEnSegundos;
     }
 }
